+++ fourier_feature_nets/__init__.py
@@ -10,9 +10,10 @@ from .fourier_feature_models import (
 )
 from .nerf_model import NeRF
 from .octree import OcTree
+from .image_dataset import ImageDataset
 from .pixel_dataset import PixelDataset
 from .ray_caster import Raycaster
-from .ray_dataset import RayData, RayDataset
+from .ray_dataset import RayDataset
 from .ray_sampler import RaySampler, RaySamples
 from .signal_dataset import SignalDataset
 from .utils import ETABar, interpolate_bilinear, load_model, orbit
@@ -29,6 +30,7 @@ __all__ = ["__version__",
            "FourierFeatureMLP",
            "PositionalFourierMLP",
            "GaussianFourierMLP",
+           "ImageDataset",
            "Voxels",
            "interpolate_bilinear",
            "load_model",
@@ -38,7 +40,6 @@ __all__ = ["__version__",
            "Raycaster",
            "RaySampler",
            "RaySamples",
-           "RayData",
            "RayDataset",
            "Resolution",
            "SignalDataset",
